Only GitLab repository browser test
