--- src/main/resources/org/jenkinsci/test/acceptance/docker/fixtures/GitListContainer/config.ini
@@ -1,27 +1,0 @@
-[git]
-client = '/usr/bin/git' ; Your git executable path
-default_branch = 'master' ; Default branch when HEAD is detached
-repositories[] = '/home/git/' ; Path to your repositories
-                                           ; If you wish to add more repositories, just add a new line
-
-; WINDOWS USERS
-;client = '"C:\Program Files (x86)\Git\bin\git.exe"' ; Your git executable path
-;repositories[] = 'C:\Path\to\Repos\' ; Path to your repositories
-
-; You can hide repositories from GitList, just copy this for each repository you want to hide
-; hidden[] = '/home/git/repositories/BetaTest'
-
-[app]
-debug = false
-cache = true
-
-; If you need to specify custom filetypes for certain extensions, do this here
-[filetypes]
-; extension = type
-; dist = xml
-
-; If you need to set file types as binary or not, do this here
-[binary_filetypes]
-; extension = true
-; svh = false
-; map = true
