--- src/main/resources/org/jenkinsci/test/acceptance/docker/fixtures/GitListContainer/nginx.conf
@@ -1,60 +1,0 @@
-user www-data;
-daemon off;
-events {
-    worker_connections  1024;
-}
-http {
-  include /etc/nginx/mime.types;
-  types_hash_max_size 2048;
-  server_names_hash_bucket_size 64;
-  server {
-      listen 80;
-      access_log /var/log/nginx/MYSERVER.access.log combined;
-      error_log /var/log/nginx/MYSERVER.error.log error;
-
-      root /var/www/gitlist;
-      index index.php;
-
-#   auth_basic "Restricted";
-#   auth_basic_user_file .htpasswd;
-
-      location = /robots.txt {
-          allow all;
-          log_not_found off;
-          access_log off;
-      }
-
-      location ~* ^/index.php.*$ {
-          fastcgi_index index.php;
-          fastcgi_param SCRIPT_FILENAME $document_root$fastcgi_script_name;
-
-          # if you're using php5-fpm via tcp
-          # fastcgi_pass 127.0.0.1:9000;
-
-          # if you're using php5-fpm via socket
-          fastcgi_pass unix:/var/run/php8.1-fpm.sock;
-
-          include /etc/nginx/fastcgi_params;
-      }
-
-      location / {
-          try_files $uri @gitlist;
-      }
-
-      location ~* \.(js|css|png|jpg|jpeg|gif|ico)$ {
-      add_header Vary "Accept-Encoding";
-          expires max;
-          try_files $uri @gitlist;
-          tcp_nodelay off;
-          tcp_nopush on;
-      }
-
-#   location ~* \.(git|svn|patch|htaccess|log|route|plist|inc|json|pl|po|sh|ini|sample|kdev4)$ {
-#       deny all;
-#   }
-
-      location @gitlist {
-          rewrite ^/.*$ /index.php;
-      }
-  }
-}
